requirements.txt
pyrofork
pillow
google-generativeai